Refactor HTTP server structure and move event handlers to v1 API namespace
