+++ hw12_13_14_15_calendar/internal/api/calendar/v1/http/handler_hello.go
@@ -1,4 +1,4 @@
-package calendar
+package http
 
 import "net/http"
